Refact assemble type (#1)
* add kms/local

* add recordable-tls

* add prover

* complete prover lib

* modify abstraction of proof

* complete listener

* add input builder

* add sp1 guest

* rename file

* add submiter

* use config

* complete test segment

* add listener test

* update listener

+++ core/src/guest.rs
@@ -31,8 +31,8 @@ pub struct GuestOutput {
 
 #[derive(Debug, Clone, Default, Serialize, Deserialize)]
 pub struct FilteredResponse {
-    pub begin: usize,
-    pub length: usize,
+    pub begin: u64,
+    pub length: u64,
 
     pub content: Vec<u8>,
 }
